src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
-e git+https://github.com/AASHE/django-integration-settings.git#egg=django-integration-settings
-e git+https://github.com/AASHE/django-s3-folder-storage.git#egg=django-s3-folder-storage
appdirs==1.4.3            # via zeep
argparse==1.4.0           # via django-htmlmin
beautifulsoup4==4.5.3     # via django-htmlmin
boto3==1.4.3
botocore==1.4.93          # via boto3, s3transfer
cached-property==1.3.0    # via zeep
click==6.6                # via pip-tools
contextlib2==0.5.4        # via raven
cssselect==1.0.3          # via premailer
cssutils==1.0.2           # via premailer
defusedxml==0.5.0         # via zeep
dj-database-url==0.5.0
django-bootstrap-pagination==1.5.0  # via django-bulletin
django-bootstrap3==7.0.0  # via django-bulletin
django-braces==1.4.0      # via django-bulletin
django-bulletin==3.4.11
django-constant-contact==1.5  # via django-bulletin
django-cors-headers==1.1.0  # via django-bulletin
django-datetime-widget==0.9.3  # via django-bulletin
django-extensions==1.7.5
django-form-utils==1.0.2  # via django-bulletin
django-haystack==2.5.1    # via django-bulletin
django-htmlmin==0.9.1
django-jsonfield==0.9.15  # via django-bulletin
django-mathfilters==0.3.0  # via django-bulletin
django-membersuite-auth==2.0
django-overextends==0.4.2
django-polymorphic==1.1   # via django-bulletin
django-positions==0.5.1   # via django-bulletin
django-storages==1.5.1
django-template-repl==0.3.0
django==1.9.13
djangorestframework==3.3.1  # via django-bulletin
docutils==0.13.1          # via botocore
elasticsearch==5.5.3
feedparser==5.2.1         # via python-constantcontact
first==2.0.1              # via pip-tools
future==0.16.0            # via django-membersuite-auth, membersuite-api-client
futures==3.0.5            # via s3transfer
gunicorn==19.6.0
html5lib==0.999999999     # via django-htmlmin
httplib2==0.11.3          # via python-constantcontact
isodate==0.5.4            # via zeep
jmespath==0.9.0           # via boto3, botocore
jsonfield==2.0.2          # via django-constant-contact
lxml==3.7.1               # via premailer, zeep
membersuite-api-client==0.4.4  # via django-membersuite-auth
nap==2.0.0                # via django-constant-contact
newrelic==2.78.0.57
olefile==0.45.1           # via pillow
pillow==4.3.0             # via django-bulletin
pip-tools==1.8.0
premailer==3.0.1          # via django-constant-contact
psycopg2==2.7.3.2
python-constantcontact==0.1  # via django-bulletin
python-dateutil==2.6.0    # via botocore
pytz==2014.7              # via django-bulletin, django-datetime-widget, zeep
raven==5.32.0
requests-toolbelt==0.7.1  # via zeep
requests==2.9.1           # via django-constant-contact, nap, premailer, requests-toolbelt, zeep
retrying==1.3.3           # via membersuite-api-client
s3transfer==0.1.10        # via boto3
six==1.10.0               # via django-braces, django-extensions, html5lib, pip-tools, python-dateutil, retrying, zeep
sorl-thumbnail==12.3      # via django-bulletin
sphinx-me==0.3            # via django-overextends
urllib3==1.24.1           # via elasticsearch
webencodings==0.5         # via html5lib
whitenoise==3.2.3
wsgiref==0.1.2            # via django-bulletin
zeep==1.2.0               # via membersuite-api-client
